+++ items/active/unsorted/HarvesterBeam/1/HarvesterBeam.activeitem
@@ -1,27 +1,27 @@
 {
-  "itemName" : "HarvesterBeam",
-  "maxStack" : 1,
-  "price" : 2500,
-  "rarity" : "uncommon",
-  "description" : "Left-click: quick harvest of crops/flowers/trees.
+	"itemName" : "HarvesterBeam",
+	"maxStack" : 1,
+	"price" : 2500,
+	"rarity" : "uncommon",
+	"description" : "Left-click: quick harvest of crops/flowers/trees.
 Right-click: till soil",
-  "shortdescription" : "Harvester Beamgun",
-  "tooltipKind" : "base",
-  "twoHanded" : true,
-"category":"tool",
-//  "learnBlueprintsOnPickup" : [ "HarvesterBeam2", "fu_autoBeamer" ],
-  "inventoryIcon" : "HarvesterBeamIcon.png",
+	"shortdescription" : "Harvester Beamgun",
+	"tooltipKind" : "base",
+	"twoHanded" : true,
+	"category":"tool",
+	//	"learnBlueprintsOnPickup" : [ "HarvesterBeam2", "fu_autoBeamer" ],
+	"inventoryIcon" : "HarvesterBeamIcon.png",
 
-  "animation" : "HarvesterBeam.animation",
-  "animationParts" : {
-    "beamgun" : "HarvesterBeam.png",
-    "beamgunfullbright" : "HarvesterBeamfullbright.png"
-  },
-  "animationCustom" : { },
-  "animationScripts" : [
-    "/items/active/unsorted/HarvesterBeam/beam.lua"
-  ],
-  "scripts" : ["HarvesterBeam.lua"],
+	"animation" : "HarvesterBeam.animation",
+	"animationParts" : {
+		"beamgun" : "HarvesterBeam.png",
+		"beamgunfullbright" : "HarvesterBeamfullbright.png"
+	},
+	"animationCustom" : { },
+	"animationScripts" : [
+		"/items/active/unsorted/HarvesterBeam/beam.lua"
+	],
+	"scripts" : ["/items/active/unsorted/HarvesterBeam/HarvesterBeam.lua"],
 
-  "harvestDistance" : 5
+	"harvestDistance" : 5
 }
