+++ items/active/unsorted/HarvesterBeam/2/HarvesterBeam2.activeitem
@@ -1,27 +1,27 @@
 {
-  "itemName" : "HarvesterBeam2",
-  "maxStack" : 1,
-  "price" : 5000,
-  "rarity" : "rare",
-  "description" : "Left-click: quick harvest of crops/flowers/trees.
+	"itemName" : "HarvesterBeam2",
+	"maxStack" : 1,
+	"price" : 5000,
+	"rarity" : "rare",
+	"description" : "Left-click: quick harvest of crops/flowers/trees.
 Right-click: till soil",
-  "shortdescription" : "Harvester Beamgun II",
-  "tooltipKind" : "base",
-  "twoHanded" : true,
-"category":"tool",
- // "learnBlueprintsOnPickup" : [ "HarvesterBeam3" ],
-  "inventoryIcon" : "HarvesterBeam2Icon.png",
+	"shortdescription" : "Harvester Beamgun II",
+	"tooltipKind" : "base",
+	"twoHanded" : true,
+	"category":"tool",
+	// "learnBlueprintsOnPickup" : [ "HarvesterBeam3" ],
+	"inventoryIcon" : "HarvesterBeam2Icon.png",
 
-  "animation" : "HarvesterBeam2.animation",
-  "animationParts" : {
-    "beamgun" : "HarvesterBeam2.png",
-    "beamgunfullbright" : "HarvesterBeam2fullbright.png"
-  },
-  "animationCustom" : { },
-  "animationScripts" : [
-    "/items/active/unsorted/HarvesterBeam/beam.lua"
-  ],
-  "scripts" : ["HarvesterBeam.lua"],
+	"animation" : "HarvesterBeam2.animation",
+	"animationParts" : {
+		"beamgun" : "HarvesterBeam2.png",
+		"beamgunfullbright" : "HarvesterBeam2fullbright.png"
+	},
+	"animationCustom" : { },
+	"animationScripts" : [
+		"/items/active/unsorted/HarvesterBeam/beam.lua"
+	],
+	"scripts" : ["/items/active/unsorted/HarvesterBeam/HarvesterBeam.lua"],
 
-  "harvestDistance" : 12
+	"harvestDistance" : 12
 }
