+++ items/active/unsorted/HarvesterBeam/3/HarvesterBeam3.activeitem
@@ -1,27 +1,27 @@
 {
-  "itemName" : "HarvesterBeam3",
-  "maxStack" : 1,
-  "price" : 7500,
-  "rarity" : "legendary",
-  "description" : "Left-click: quick harvest of crops/flowers/trees.
+	"itemName" : "HarvesterBeam3",
+	"maxStack" : 1,
+	"price" : 7500,
+	"rarity" : "legendary",
+	"description" : "Left-click: quick harvest of crops/flowers/trees.
 Right-click: till soil",
-  "shortdescription" : "Harvester Beamgun III",
-  "tooltipKind" : "base",
-  "twoHanded" : true,
-"category":"tool",
+	"shortdescription" : "Harvester Beamgun III",
+	"tooltipKind" : "base",
+	"twoHanded" : true,
+	"category":"tool",
 
-  "inventoryIcon" : "HarvesterBeam3Icon.png",
+	"inventoryIcon" : "HarvesterBeam3Icon.png",
 
-  "animation" : "HarvesterBeam3.animation",
-  "animationParts" : {
-    "beamgun" : "HarvesterBeam3.png",
-    "beamgunfullbright" : "HarvesterBeam3fullbright.png"
-  },
-  "animationCustom" : { },
-  "animationScripts" : [
-    "/items/active/unsorted/HarvesterBeam/beam.lua"
-  ],
-  "scripts" : ["HarvesterBeam.lua"],
+	"animation" : "HarvesterBeam3.animation",
+	"animationParts" : {
+		"beamgun" : "HarvesterBeam3.png",
+		"beamgunfullbright" : "HarvesterBeam3fullbright.png"
+	},
+	"animationCustom" : { },
+	"animationScripts" : [
+		"/items/active/unsorted/HarvesterBeam/beam.lua"
+	],
+	"scripts" : ["/items/active/unsorted/HarvesterBeam/HarvesterBeam.lua"],
 
-  "harvestDistance" : 24
+	"harvestDistance" : 24
 }
