+++ items/active/unsorted/HarvesterBeam/beam.lua
@@ -4,6 +4,6 @@ require "/scripts/util.lua"
 function update()
 	localAnimator.clearDrawables()
     for _,l in pairs(animationConfig.animationParameter("lineList", {})) do
-	  localAnimator.addDrawable({line = l.line, width = l.width, color = l.color, position = l.position, fullbright = true})
+		localAnimator.addDrawable({line = l.line, width = l.width, color = l.color, position = l.position, fullbright = true})
     end
 end
